requirements.txt
torch
torchvision
timm
numpy
Pillow
einops
swanlab
